+++ src/security/vcc_pki_client.cpp
@@ -427,7 +427,7 @@ X509Certificate VCCPKIClient::parseCertificate(const std::string& pem) {
     X509Certificate cert;
     
     // Parse PEM using OpenSSL
-    BIO_ptr bio(BIO_new_mem_buf(pem.c_str(), pem.size()));
+    BIO_ptr bio(BIO_new_mem_buf(pem.c_str(), static_cast<int>(pem.size())));
     if (!bio) {
         throw std::runtime_error("Failed to create BIO from PEM");
     }
@@ -487,7 +487,7 @@ X509Certificate VCCPKIClient::parseCertificate(const std::string& pem) {
 
 bool VCCPKIClient::validateCertChain(const X509Certificate& cert) const {
     // Load the certificate from PEM
-    BIO_ptr cert_bio(BIO_new_mem_buf(cert.pem.data(), cert.pem.size()));
+    BIO_ptr cert_bio(BIO_new_mem_buf(cert.pem.data(), static_cast<int>(cert.pem.size())));
     if (!cert_bio) {
         return false;
     }
